fix(tunnel): oblique view, drifting mouth, and no more stutter
Three faults, all visible on the panel and all now fixed against the
reference capture:

- Straight-on. The tunnel axis now sits well off the field's centre, so the
  mouth rides high and the near wall stretches away below. A tunnel viewed
  down its own axis is a symmetric target, which is the one thing the
  reference never looks like.
- Static. The depth and angle fields are computed LARGER than the screen
  (152x152 for a 120x120 window) and the window pans across them. Panning
  the window IS moving the camera off-axis, because both coordinates depend
  only on the offset from the tunnel axis — so the vanishing point wanders
  without recomputing a single atan2.
- Jitter. The roll advanced 0.39 texels per frame, so the integer shift sat
  still for two frames and then jumped one, and at the rim a texel was ~8
  screen pixels. Both shifts now advance exactly one texel per frame, and the
  angular axis widened to 256 texels so the step is small and the speed still
  right — 32 cells per revolution, which also matches the reference.

The pan is what costs: the pool grows to 61 KB, free heap 76.6 -> 59.4 KB,
steady. TUN_MARGIN is the dial — ~1 KB per half-res pixel of travel, and zero
margin puts the pool back at 46 KB.

+++ src/app/effects/TunnelEffect.h
@@ -10,9 +10,9 @@
 #if SMOLBASE_FRAMEBUFFER == SMOLBASE_FB_PALETTE_8
 
 class TunnelEffect : public Effect {
-  uint16_t uAcc = 0;  // 8.8 fixed: depth shift, the fall down the pipe
-  uint16_t vAcc = 0;  // 8.8 fixed: angle shift, the roll around it
-  uint16_t pulse = 0; // the slow breath in the ramp's hot end
+  uint8_t uAcc = 0;   // depth shift: the fall down the pipe, one texel a frame
+  uint8_t vAcc = 0;   // angle shift: the roll around it, one texel a frame
+  uint16_t pulse = 0; // frame counter: the ramp's breath and the window's pan
   int8_t dir = 1;     // tap flips it: fly in, or fall out
 
   void buildTexture();
